refactor(id's): Update the definition of id's so they are int and not int64 given that in the sql declaration they were int and not big ints

+++ app/internal/image/model.go
@@ -5,7 +5,7 @@ import "github.com/uptrace/bun"
 type Image struct {
 	bun.BaseModel `bun:"table:image,alias:img" swaggerignore:"true"`
 
-	ImageID int64  `bun:"image_id,pk,autoincrement" json:"image_id"`
+	ImageID int    `bun:"image_id,pk,autoincrement" json:"image_id"`
 	Name    string `bun:"name"                      json:"name"`
 	URL     string `bun:"url"                       json:"url"`
 }
